+++ 2-library-tasks/task-4/src/main/java/ru/otus/spring/shepin/dao/GenreRepository.java
@@ -3,6 +3,6 @@ package ru.otus.spring.shepin.dao;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ru.otus.spring.shepin.entity.Genre;
 
-public interface GenreRepository extends MongoRepository<Genre, Long> {
+public interface GenreRepository extends MongoRepository<Genre, Integer> {
     Genre getByName(String name);
 }
