+++ 2-library-tasks/task-4/src/main/java/ru/otus/spring/shepin/entity/Comment.java
@@ -1,8 +1,12 @@
 package ru.otus.spring.shepin.entity;
 
 
-
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -15,7 +19,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @Document(collation ="comment")
 public class Comment {
     @Id
-    private Long id;
+    private int id;
 
     String commentText;
