+++ 2-library-tasks/task-4/src/main/java/ru/otus/spring/shepin/service/AuthorService.java
@@ -9,5 +9,5 @@ public interface AuthorService {
 
     Author create(String firstName, String lastName);
 
-    Author getById(Long id);
+    Author getById(Integer id);
 }
