+++ 2-library-tasks/task-4/src/main/java/ru/otus/spring/shepin/service/CommentService.java
@@ -5,7 +5,7 @@ import ru.otus.spring.shepin.entity.Comment;
 import java.util.List;
 
 public interface CommentService {
-    Comment createByParams(Long bookId, String lastName);
-    public List<Comment> getAllCommentsByBookId(Long bookId);
-    void deleteAllCommentsByBookId(Long bookId);
+    Comment createByParams(Integer bookId, String lastName);
+    public List<Comment> getAllCommentsByBookId(Integer bookId);
+    void deleteAllCommentsByBookId(Integer bookId);
 }
